requirements.txt
neuralgym==0.0.1
numpy==1.24.3
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.1.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
tqdm==4.66.1
